requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
cryptography>=41.0.0
python-dotenv>=1.0.0